get rid of gettickcount

+++ GUI/host/textthread.h
@@ -36,7 +36,7 @@ private:
 	BYTE leadByte = 0;
 	std::unordered_set<wchar_t> repeatingChars;
 	std::mutex bufferMutex;
-	DWORD lastPushTime = 0;
+	DWORD64 lastPushTime = 0;
 	Synchronized<std::vector<std::wstring>> queuedSentences;
 	struct TimerDeleter { void operator()(HANDLE h) { DeleteTimerQueueTimer(NULL, h, INVALID_HANDLE_VALUE); } };
 	AutoHandle<TimerDeleter> timer = NULL;
